helper-apps/cortex-claude-code-agent-sdk/requirements.txt
claude-agent-sdk>=0.1.0
aiohttp>=3.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
